# api/docs/swagger.yaml
definitions:
  models.ForbiddenError:
    properties:
      message:
        type: string
    type: object
  models.StandartError:
    properties:
      error: {}
    type: object
  models.Task:
    properties:
      description:
        type: string
      done:
        type: boolean
      id:
        type: integer
      title:
        type: string
    type: object
info:
  contact: {}
  description: This swagger UI was created in lesson
  title: 'Project: Swagger Intro'
  version: "1.0"
paths:
  /tasks:
    delete:
      consumes:
      - application/json
      description: This method delete tasks
      parameters:
      - description: ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandartError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ForbiddenError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandartError'
      security:
      - BearerAuth: []
      summary: delte TASKS
      tags:
      - TASK Tag
    get:
      description: This method gets tasks
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/models.Task'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandartError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ForbiddenError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandartError'
      security:
      - BearerAuth: []
      summary: GET TASKS
      tags:
      - TASK Tag
    post:
      consumes:
      - application/json
      description: This method creates tasks
      parameters:
      - description: Task
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Task'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandartError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ForbiddenError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandartError'
      security:
      - BearerAuth: []
      summary: CREATE TASKS
      tags:
      - TASK
    put:
      consumes:
      - application/json
      description: This method put tasks
      parameters:
      - description: ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandartError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ForbiddenError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandartError'
      security:
      - BearerAuth: []
      summary: put TASKS
      tags:
      - TASK Tag
securityDefinitions:
  Bearerauth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"